% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rfun_randomGraph.R
\name{randomGraph}
\alias{randomGraph}
\title{Generate a Random Graph}
\usage{
randomGraph(n, seed = as.numeric(Sys.time()), wlim = c(0, 1), Glim = c(0, 1))
}
\arguments{
\item{n}{an integer: number of vertices}

\item{seed}{an integer: a seed for random number generator}

\item{wlim}{a vector of two numbers: range of vertex weights}

\item{Glim}{a vector of two numbers: range of transition weights}
}
\value{
A list of one vector for vertex weights and one matrix for transition weights
}
\description{
Generate a random graph from uniform distribution
}
\examples{
wG <- randomGraph(n=5)
}
\author{
Jiangtao Gou

Fengqing Zhang
}
